resolve cyclic dependencies
