+++ pallets/asset/types/src/lib.rs
@@ -1,0 +1,194 @@
+use codec::{Decode, Encode};
+use polymesh_primitives::{IdentityId, Ticker};
+use polymesh_primitives_derive::VecU8StrongTyped;
+#[cfg(feature = "std")]
+use serde::{Deserialize, Serialize};
+use sp_std::prelude::*;
+
+/// The type of an asset represented by a token.
+#[derive(Encode, Decode, Clone, Debug, PartialEq, Eq)]
+pub enum AssetType {
+    EquityCommon,
+    EquityPreferred,
+    Commodity,
+    FixedIncome,
+    REIT,
+    Fund,
+    RevenueShareAgreement,
+    StructuredProduct,
+    Derivative,
+    Custom(Vec<u8>),
+}
+
+impl Default for AssetType {
+    fn default() -> Self {
+        AssetType::Custom(b"undefined".to_vec())
+    }
+}
+
+/// The type of an identifier associated with a token.
+// these are used off-chain.
+#[derive(Encode, Decode, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
+pub enum IdentifierType {
+    Cins,
+    Cusip,
+    Isin,
+    Dti,
+}
+
+impl Default for IdentifierType {
+    fn default() -> Self {
+        IdentifierType::Isin
+    }
+}
+
+/// Ownership status of a ticker/token.
+#[derive(Encode, Decode, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
+pub enum AssetOwnershipRelation {
+    NotOwned,
+    TickerOwned,
+    AssetOwned,
+}
+
+impl Default for AssetOwnershipRelation {
+    fn default() -> Self {
+        Self::NotOwned
+    }
+}
+
+/// A wrapper for a token name.
+#[derive(
+    Decode, Encode, Clone, Debug, Default, Hash, PartialEq, Eq, PartialOrd, Ord, VecU8StrongTyped,
+)]
+pub struct AssetName(pub Vec<u8>);
+
+/// A wrapper for an asset ID.
+#[derive(
+    Decode, Encode, Clone, Debug, Default, Hash, PartialEq, Eq, PartialOrd, Ord, VecU8StrongTyped,
+)]
+pub struct AssetIdentifier(pub Vec<u8>);
+
+/// A wrapper for a funding round name.
+#[derive(Decode, Encode, Clone, Debug, Hash, PartialEq, Eq, PartialOrd, Ord, VecU8StrongTyped)]
+pub struct FundingRoundName(pub Vec<u8>);
+
+impl Default for FundingRoundName {
+    fn default() -> Self {
+        FundingRoundName("".as_bytes().to_vec())
+    }
+}
+
+/// struct to store the token details.
+#[derive(Encode, Decode, Default, Clone, PartialEq, Debug)]
+pub struct SecurityToken<U> {
+    pub name: AssetName,
+    pub total_supply: U,
+    pub owner_did: IdentityId,
+    pub divisible: bool,
+    pub asset_type: AssetType,
+    pub primary_issuance_agent: Option<IdentityId>,
+}
+
+/// struct to store the signed data.
+#[derive(Encode, Decode, Default, Clone, PartialEq, Debug)]
+pub struct SignData<U> {
+    pub custodian_did: IdentityId,
+    pub holder_did: IdentityId,
+    pub ticker: Ticker,
+    pub value: U,
+    pub nonce: u16,
+}
+
+/// struct to store the ticker registration details.
+#[derive(Encode, Decode, Clone, Default, PartialEq, Debug)]
+pub struct TickerRegistration<U> {
+    pub owner: IdentityId,
+    pub expiry: Option<U>,
+}
+
+/// struct to store the ticker registration config.
+#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
+#[derive(Encode, Decode, Clone, Default, PartialEq, Debug)]
+pub struct TickerRegistrationConfig<U> {
+    pub max_ticker_length: u8,
+    pub registration_length: Option<U>,
+}
+
+/// Enum that represents the current status of a ticker.
+#[derive(Encode, Decode, Clone, Eq, PartialEq, Debug)]
+pub enum TickerRegistrationStatus {
+    RegisteredByOther,
+    Available,
+    RegisteredByDid,
+}
+
+/// Enum that uses as the return type for the restriction verification.
+#[derive(Encode, Decode, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
+pub enum RestrictionResult {
+    Valid,
+    Invalid,
+    ForceValid,
+}
+
+impl Default for RestrictionResult {
+    fn default() -> Self {
+        RestrictionResult::Invalid
+    }
+}
+
+/// The total asset balance and the balance of the asset in a specified portfolio of an identity.
+#[cfg_attr(feature = "std", derive(Debug))]
+#[derive(Clone, PartialEq, Eq, PartialOrd, Ord)]
+pub struct FocusedBalances<Balance> {
+    /// The total balance of the asset held by the identity.
+    pub total: Balance,
+    /// The balance of the asset in the default portfolio of the identity.
+    pub portfolio: Balance,
+}
+
+/// An Ethereum address (i.e. 20 bytes, used to represent an Ethereum account).
+///
+/// This gets serialized to the 0x-prefixed hex representation.
+#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
+#[derive(Clone, Copy, PartialEq, Eq, Encode, Decode, Default, Debug)]
+pub struct EthereumAddress(pub [u8; 20]);
+
+#[derive(Encode, Decode, Clone)]
+pub struct EcdsaSignature(pub [u8; 65]);
+
+impl PartialEq for EcdsaSignature {
+    fn eq(&self, other: &Self) -> bool {
+        &self.0[..] == &other.0[..]
+    }
+}
+
+impl sp_std::fmt::Debug for EcdsaSignature {
+    fn fmt(&self, f: &mut sp_std::fmt::Formatter<'_>) -> sp_std::fmt::Result {
+        write!(f, "EcdsaSignature({:?})", &self.0[..])
+    }
+}
+
+/// Data imported from Polymath Classic regarding ticker registration/creation.
+/// Only used at genesis config and not stored on-chain.
+#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
+#[derive(Clone)]
+pub struct ClassicTickerImport {
+    /// Owner of the registration.
+    pub eth_owner: EthereumAddress,
+    /// Name of the ticker registered.
+    pub ticker: Ticker,
+    /// Is `eth_owner` an Ethereum contract (e.g., in case of a multisig)?
+    pub is_contract: bool,
+    /// Has the ticker been elevated to a created asset on classic?
+    pub is_created: bool,
+}
+
+/// Data about a ticker registration from Polymath Classic on-genesis importation.
+#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
+#[derive(Encode, Decode, Clone, Debug, PartialEq, Eq)]
+pub struct ClassicTickerRegistration {
+    /// Owner of the registration.
+    pub eth_owner: EthereumAddress,
+    /// Has the ticker been elevated to a created asset on classic?
+    pub is_created: bool,
+}
