+++ pallets/common/Cargo.toml
@@ -8,7 +8,7 @@ edition = "2018"
 # Common
 polymesh-primitives = { path = "../../primitives", default-features = false }
 polymesh-primitives-derive = { path = "../../primitives_derive", default-features = false }
-pallet-asset = { path = "../asset", default-features = false }
+pallet-asset-types = { path = "../asset/types", default-features = false }
 
 # Other
 serde = { version = "1.0.112", default-features = false }
